requirements.txt
accelerate==0.29.3
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
distro==1.9.0
fastapi==0.110.2
filelock==3.13.4
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
isort==5.13.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.24.0
packaging==24.0
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
safetensors==0.4.3
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tokenizers==0.19.1
tomlkit==0.12.4
torch==2.3.0
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
